float64

+++ tests/hip/dot.nim
@@ -10,23 +10,23 @@ const
 # TODO improve this
 {.pragma: hippoShared, exportc, codegenDecl: "__shared__ $# $#".}
 
-proc dot(a, b, c: ptr[cfloat]){.hippoGlobal.} =
+proc dot(a, b, c: ptr[float64]){.hippoGlobal.} =
   
-  #var cache {.hippoShared.}: ptr[cfloat]
-  #var cache {.importc: "cache"}: ptr[cfloat]
+  #var cache {.hippoShared.}: ptr[float64]
+  #var cache {.importc: "cache"}: ptr[float64]
   # __shared__ float cache[threadsPerBlock
   {.emit:["""__shared__ float cache[256];"""].}
-  let cache {.codeGenDecl:"" importc noinit.}: ptr UncheckedArray[cfloat]
-  #var cache {.hippoShared.}: ptr[cfloat]
+  let cache {.codeGenDecl:"" importc noinit.}: ptr UncheckedArray[float64]
+  #var cache {.hippoShared.}: ptr[float64]
 
   # TODO figure out how to do this properly
-  let aArray = cast[ptr UncheckedArray[cfloat]](a)
-  let bArray = cast[ptr UncheckedArray[cfloat]](b)
-  let cArray = cast[ptr UncheckedArray[cfloat]](c)
+  let aArray = cast[ptr UncheckedArray[float64]](a)
+  let bArray = cast[ptr UncheckedArray[float64]](b)
+  let cArray = cast[ptr UncheckedArray[float64]](c)
 
   let cacheIndex = threadIdx.x
   var tid = threadIdx.x + blockIdx.x * blockDim.x
-  var temp: cfloat = 0
+  var temp: float64 = 0
   while tid < N:
     # TODO not sure how to handle functions like `pluseq___dot_u17` with nim / cuda
     temp = temp + (aArray[tid] * bArray[tid])
@@ -53,13 +53,13 @@ proc dot(a, b, c: ptr[cfloat]){.hippoGlobal.} =
 
 
 proc main() =
-  var a, b, partial_c: array[N, cfloat]
+  var a, b, partial_c: array[N, float64]
   var dev_a, dev_b, dev_partial_c: pointer
 
   # allocate gpu memory
-  handleError(hipMalloc(addr dev_a, sizeof(cfloat)*N))
-  handleError(hipMalloc(addr dev_b, sizeof(cfloat)*N))
-  handleError(hipMalloc(addr dev_partial_c, BlocksPerGrid * sizeof(cfloat)))
+  handleError(hipMalloc(addr dev_a, sizeof(float64)*N))
+  handleError(hipMalloc(addr dev_b, sizeof(float64)*N))
+  handleError(hipMalloc(addr dev_partial_c, BlocksPerGrid * sizeof(float64)))
 
   # fill in host memory with data
   for i in 0 ..< N:
@@ -67,8 +67,8 @@ proc main() =
     b[i] = (i * 2).float
 
   # copy data to device
-  handleError(hipMemcpy(dev_a, addr a[0], sizeof(cfloat)*N, hipMemcpyHostToDevice))
-  handleError(hipMemcpy(dev_b, addr b[0], sizeof(cfloat)*N, hipMemcpyHostToDevice))
+  handleError(hipMemcpy(dev_a, addr a[0], sizeof(float64)*N, hipMemcpyHostToDevice))
+  handleError(hipMemcpy(dev_b, addr b[0], sizeof(float64)*N, hipMemcpyHostToDevice))
 
   # launch kernel
   handleError(launchKernel(
@@ -79,14 +79,14 @@ proc main() =
   ))
 
   # copy memory back from GPU to CPU
-  handleError(hipMemcpy(addr partial_c[0], dev_partial_c, BlocksPerGrid * sizeof(cfloat), hipMemcpyDeviceToHost))
+  handleError(hipMemcpy(addr partial_c[0], dev_partial_c, BlocksPerGrid * sizeof(float64), hipMemcpyDeviceToHost))
 
   # finish up on the CPU
-  var c: cfloat = 0
+  var c: float64 = 0
   for i in 0 ..< BlocksPerGrid:
     c += partial_c[i]
 
-  proc sum_squares(x: cfloat): cfloat =
+  proc sum_squares(x: float64): float64 =
     result = x * (x + 1) * (2 * x + 1) / 6
   
   echo fmt"Does GPU value {c:e} = {2 * sum_squares((N - 1)):e}?"
